+++ src/StructuredField.php
@@ -7,8 +7,8 @@ namespace Bakame\Http\StructuredFields;
 use Stringable;
 
 /**
- * @phpstan-type SfType ByteSequence|Token|\DateTimeInterface|string|int|float|bool
- * @phpstan-type SfTypeInput StructuredField|SfType
+ * @phpstan-type SfType ByteSequence|Token|\DateTimeImmutable|string|int|float|bool
+ * @phpstan-type SfTypeInput StructuredField|SfType|\DateTimeInterface
  * @phpstan-type SfItem ValueAccess&ParameterAccess
  * @phpstan-type SfItemInput SfItem|SfTypeInput
  * @phpstan-type SfMember (MemberList<int, SfItem>|ValueAccess)&ParameterAccess